refactor(tests): split dispatcher auto-compat tests
